+++ app/Http/Controllers/CategoryJudgeController.php
@@ -6,15 +6,15 @@ use Illuminate\Http\Request;
 use App\Contest;
 use App\ContestCategory;
 use App\CategoryJudge;
-use App\Judge;
+use App\User;
 
 class CategoryJudgeController extends Controller
 {
     public function __construct()
     {
-		$this->middleware('auth');
-		
-		$this->middlware('adminUser');
+        $this->middleware('auth');
+        
+        $this->middleware('adminUser');
     }
 
     /**
@@ -23,11 +23,15 @@ class CategoryJudgeController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request, Contest $contest, ContestCategory $contestCategory, Judge $judge)
+    public function store(Request $request, Contest $contest, ContestCategory $contestCategory, User $judge)
     {
+        if ($contestCategory->status != 'que') {
+            return redirect('/contests/' . $contest->id . '/categories/' . $contestCategory->id . '?activeTab=Judges')->with('error', 'Could not Add Judge. Please make sure that the Category has not Started Scoring or Finished Scoring.');
+        }
+
         $data = [
             'contest_category_id' => $contestCategory->id,
-            'judge_id' => $judge->id
+            'user_id' => $judge->id
         ];
 
         CategoryJudge::create($data);
@@ -43,6 +47,10 @@ class CategoryJudgeController extends Controller
      */
     public function destroy(Contest $contest, ContestCategory $contestCategory, CategoryJudge $categoryJudge)
     {
+        if ($contestCategory->status != 'que') {
+            return redirect('/contests/' . $contest->id . '/categories/' . $contestCategory->id . '?activeTab=Judges')->with('error', 'Could not Remove Judge. Please make sure that the Category has not Started Scoring or Finished Scoring.');
+        }
+
         $categoryJudge->delete();
 
         return redirect('/contests/' . $contest->id . '/categories/' . $contestCategory->id . '?activeTab=Judges')->with('success', 'Judge has been Removed.');
